Expose glTF API as function call
