--- modules/core/src/io/index.js
@@ -1,2 +1,0 @@
-// Call a require based helper to select platform to export
-export {setPathPrefix, loadFile, loadImage} from './browser-load';
