+++ modules/core/src/scenegraph/wip/mesh.js
